+++ src/hal/drivers/mesa-hostmot2/llio_info.c
@@ -224,7 +224,7 @@ const char *set_llio_info_spi(hm2_lowlevel_io_t *llio, const hm2_idrom_t *idrom)
 
 	/* In the far future, when there are too many boards, use bsearch */
 	/* With few boards, linear search is faster */
-	for(i = 0; i < NELEM(spiboards); i++) {
+	for(i = 0; (size_t)i < NELEM(spiboards); i++) {
 		if(!memcmp(idrom->board_name, spiboards[i].board_name, sizeof(idrom->board_name))) {
 			llio->num_ioport_connectors = spiboards[i].num_ioport_connectors;
 			llio->pins_per_connector = spiboards[i].pins_per_connector;
@@ -245,7 +245,7 @@ const char *set_llio_info_spi(hm2_lowlevel_io_t *llio, const hm2_idrom_t *idrom)
 
 	memcpy(buf, idrom->board_name, sizeof(idrom->board_name));
 	buf[sizeof(idrom->board_name)] = 0;
-	for(i = 0; i < sizeof(idrom->board_name); i++) {
+	for(i = 0; (size_t)i < sizeof(idrom->board_name); i++) {
 		if(!isprint(buf[i]))
 			buf[i] = '?';
 	}
