+++ src/hal/drivers/mesa-hostmot2/ssr.c
@@ -96,7 +96,7 @@ int hm2_ssr_parse_md(hostmot2_t *hm2, int md_index) {
             int out;
             for (
                 out = 0;
-                out < sizeof(hm2->ssr.instance[0].hal.pin.out)/sizeof(hm2->ssr.instance[0].hal.pin.out[0]);
+                (size_t)out < sizeof(hm2->ssr.instance[0].hal.pin.out)/sizeof(hm2->ssr.instance[0].hal.pin.out[0]);
                 out ++
             ) {
                 hm2->ssr.instance[inst].hal.pin.out[out] = NULL;
